+++ packages/ocean-react/rollup.config.mjs
@@ -1,9 +1,10 @@
-import resolve from '@rollup/plugin-node-resolve';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import typescript from 'rollup-plugin-typescript2';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
+import { readFileSync } from 'fs';
 
-import pkg from './package.json';
+const pkg = JSON.parse(readFileSync('./package.json', 'utf8'));
 
 export default {
   input: 'src/index.ts',
@@ -21,7 +22,7 @@ export default {
   ],
   plugins: [
     peerDepsExternal(),
-    resolve(),
+    nodeResolve(),
     commonjs(),
     typescript({
       tsconfig: 'tsconfig.build.json',