+++ packages/ocean-core/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.86.0](https://github.com/ocean-ds/ocean-web/compare/v1.85.0...v1.86.0) (2025-07-21)
+
+### Features
+
+- add ocean docs docusaurus ([#1153](https://github.com/ocean-ds/ocean-web/issues/1153)) ([1de3ca1](https://github.com/ocean-ds/ocean-web/commit/1de3ca136705d14cea29a35c78dc01f397fd46f3))
+
 # [1.79.0](https://github.com/ocean-ds/ocean-web/compare/v1.78.0...v1.79.0) (2025-07-03)
 
 ### Features